Move tests to more appropriate directories
