+++ tests/src/test/java/org/mozilla/javascript/tests/es2024/DataViewResizableArrayBufferTest.java
@@ -1,10 +1,10 @@
-package org.mozilla.javascript.tests.harmony;
+package org.mozilla.javascript.tests.es2024;
 
 import org.mozilla.javascript.Context;
 import org.mozilla.javascript.drivers.LanguageVersion;
 import org.mozilla.javascript.drivers.RhinoTest;
 import org.mozilla.javascript.drivers.ScriptTestsBase;
 
-@RhinoTest("testsrc/jstests/harmony/typedarray-resizablearraybuffer.js")
+@RhinoTest("testsrc/jstests/es2024/dataview-resizablearraybuffer.js")
 @LanguageVersion(Context.VERSION_ES6)
-public class TypedArrayResizableArrayBufferTest extends ScriptTestsBase {}
+public class DataViewResizableArrayBufferTest extends ScriptTestsBase {}
