+++ tests/src/test/java/org/mozilla/javascript/tests/es2024/TypedArrayResizableArrayBufferTest.java
@@ -1,10 +1,10 @@
-package org.mozilla.javascript.tests.harmony;
+package org.mozilla.javascript.tests.es2024;
 
 import org.mozilla.javascript.Context;
 import org.mozilla.javascript.drivers.LanguageVersion;
 import org.mozilla.javascript.drivers.RhinoTest;
 import org.mozilla.javascript.drivers.ScriptTestsBase;
 
-@RhinoTest("testsrc/jstests/harmony/dataview-resizablearraybuffer.js")
+@RhinoTest("testsrc/jstests/es2024/typedarray-resizablearraybuffer.js")
 @LanguageVersion(Context.VERSION_ES6)
-public class DataViewResizableArrayBufferTest extends ScriptTestsBase {}
+public class TypedArrayResizableArrayBufferTest extends ScriptTestsBase {}
