+++ tests/testsrc/jstests/es2024/dataview-resizablearraybuffer.js
