+++ tests/testsrc/jstests/es2024/typedarray-resizablearraybuffer.js
